+++ src/routes/index.jsx
@@ -1,4 +1,4 @@
-import Route from './classes';
+import Routes from './classes';
 import News from './Newsletter';
 import Dashboard from './Dashboard';
 import Header from './Header';
@@ -6,79 +6,28 @@ import Guarantee from './Guarantee';
 import Education from './Education';
 // console.log('Header', Header);
 
-class Routes {
-    constructor(list) {
-        // super(props);
-        this.list = list;
-    }
-
-    getLinks() {
-        return this.list.reduce((acc, next) => {
-            if (!next.label) {
-                return acc;
-            }
-            return acc.concat({
-                label: next.label,
-                key: next.key,
-                to: next.path,
-            });
-        }, []);
-    }
-
-    handleComponents(isMobile) {
-        return this.list.map(item => {
-            const { component, sidebar, main, key, path, exact } = item;
-            let con, side, ma;
-            if (Array.isArray(component)) {
-                con = isMobile ? component[0] : component[1];
-            } else {
-                con = component;
-            }
-
-            if (Array.isArray(sidebar)) {
-                side = isMobile ? component[0] : component[1];
-            } else {
-                side = sidebar;
-            }
-
-            if (Array.isArray(main)) {
-                ma = isMobile ? component[0] : component[1];
-            } else {
-                ma = main;
-            }
-            return {
-                component: con,
-                sidebar: side,
-                main: ma,
-                exact,
-                path,
-                key
-            };
-        });
-    }
-}
 
 const routes = new Routes([
-    new Route({
+    {
         component: Header
-    }),
-    new Route({
+    },
+    {
         component: Dashboard,
         exact: true
-    }),
-    new Route({
+    },
+    {
         component: Guarantee,
         path: '/guarantee',
         label: 'guarantee'
-    }),
-    new Route({
+    },
+    {
         component: Education,
         path: '/education',
         label: 'education'
-    }),
-    new Route({
+    },
+    {
         component: News
-    })
+    }
 ]);
 
 // function getLinks() {
